requirements.txt
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest
pytest-base-url
pytest
pymysql
pyyaml



rsa
requests
jsonpath